% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{as.fun}
\alias{as.fun}
\title{Convert Formula to Function}
\usage{
as.fun(~ x > 3)
}
\arguments{
\item{f}{a right sided formula}
}
\value{
a function which evaluates the given formula
}
\description{
Convert Formula to Function
}
\note{
the function passes an argument x, so x should be the variable in the formula
}
