% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\title{Assert the class and length of an object}
\usage{
assert(
  x,
  what,
  name = NULL,
  length_check = TRUE,
  length_req = NULL,
  stopOnFail = NULL
)
}
\arguments{
\item{x}{an R object}

\item{what}{(character) The class(es) that the object should be}

\item{name}{(character) The name of the R object (used in error messages)}

\item{length_check}{(logical) Whether to check if the object has length (Default: TRUE)}

\item{length_req}{(numeric) The required length of x.}

\item{stopOnFail}{Whether the function stops as soon as one of the check fails, or keep going to gather all error messages at once.}
}
\description{
Assert the class and length of an object
}
