% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_boolean}
\alias{assert_boolean}
\title{Assert if an object contains only TRUE or FALSE values}
\usage{
assert_boolean(x, name = lang2str(x), is2nd = FALSE, stopOnFail = TRUE)
}
\arguments{
\item{x}{(character | numeric | integer | logical) A vector.}

\item{name}{(character) The name of the vector (used in error messages). Optional.}

\item{stopOnFail}{(logical) Whether the function stops as soon as one of the check fails, or gathers all error messages first.}
}
\description{
Check that each element of x is either TRUE or FALSE. Works with vectors containing logical values, and/or 0s and 1s, and/or "true", "t", "false", "f" strings (regardless of case).
}
