% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_gbif.R
\name{classification_gbif}
\alias{classification_gbif}
\title{Get taxonomic classification data from GBIF}
\usage{
classification_gbif(
  name,
  rank = NULL,
  kingdom = NULL,
  phylum = NULL,
  class = NULL,
  order = NULL,
  family = NULL,
  genus = NULL,
  strict = FALSE,
  start = NULL,
  asList = TRUE,
  rows = NULL
)
}
\arguments{
\item{name}{character. Scientific name of the taxon.}

\item{rank}{character. Taxonomic rank of the taxon. One of 'kingdom', 'phylum', 'class', 'order', 'family', 'genus' or 'species'.}

\item{kingdom}{character. Which kingdom should the taxon be a child of.}

\item{phylum}{character. Which phylum should the taxon be a child of. Ignored if rank is higher.}

\item{class}{character. Which class should the taxon be a child of. Ignored if rank is higher.}

\item{order}{character. Which order should the taxon be a child of. Ignored if rank is higher.}

\item{family}{character. Which family should the taxon be a child of. Ignored if rank is higher.}

\item{genus}{character. Which genus should the taxon be a child of. Ignored if rank is higher.}

\item{strict}{logical.   If matching should be strict.}

\item{start}{numeric.}

\item{asList}{logical.   If the output should be a list. Default TRUE.}

\item{rows}{numeric.}
}
\description{
Get taxonomic classification data from GBIF
}
