% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat.R
\name{concat}
\alias{concat}
\title{Concatenate elements of a vector.}
\usage{
concat(x, sep = "")
}
\arguments{
\item{x}{(character) A character vector.}

\item{sep}{(character) An character string to separate the elements.
Default is "".}
}
\value{
A character string of the concatenated values.
}
\description{
Concatenate elements of a vector after converting to character.
}
