% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortenFuns.R
\name{df}
\alias{df}
\title{shortcut for `data.frame()` allowing to name the columns as well
... same arguments as for `data.frame()`
names character vector for the names of the data.frame (colnames)}
\usage{
df(..., names = NULL)
}
\description{
shortcut for `data.frame()` allowing to name the columns as well
... same arguments as for `data.frame()`
names character vector for the names of the data.frame (colnames)
}
