% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_fill.R
\name{empty_fill}
\alias{empty_fill}
\title{Fill missing values}
\usage{
empty_fill(x, type = c("const", "locf", "nocb"), fill = NA)
}
\arguments{
\item{x}{vector, list, data.frame or data.table}

\item{type}{character, one of "const", "locf" or "nocb". Defaults to "const"}

\item{fill}{value to be used to fill when type=="const"}

\item{cols}{numeric or character vector specifying columns to be updated. description}
}
\value{
x without missing values
}
\description{
Fill missing values
}
