% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{extractting Vectors and List}
\usage{
extract(1:5, ~x>3)

            dat <- list(list(item = 1, data = NULL), list(item = 2, data = data.frame(x = letters[1:5], y = 1:5)))list(list(item = 1, data = NULL), list(item = 2, data = data.frame(x = letters[1:5], y = 1:5)))
            extract(dat, ~!is.null(x$data))
}
\arguments{
\item{x}{a data vector (including list) to be filtered}

\item{f}{a function or a formula to use as a function. Return must be of type logical.}
}
\value{
the filtered data object
}
\description{
extract extracts the elements of a vector for which a formula or function gives true. If the object is a list, the formula or function is evaluated using vapply.
}
