% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head2.R
\name{head2}
\alias{head2}
\title{Return the First or Last Parts of an Object}
\usage{
head2(x, n = 6)
}
\arguments{
\item{x}{an object}

\item{n}{an integer vector of length 1 or up to dim(x). Default is 6L. Negative values subtract n to the length or dimension. If only one value is provided but the object has more than one dimension, it will be used for all dimension (the difference with base::head()). If the object is a function, it will return n lines of the source code.}
}
\value{
A subsetted object.
}
\description{
Return the First or Last Parts of an Object
}
