% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertColumn.R
\name{insertCol}
\alias{insertCol}
\title{Insert a column into a data.frame object}
\usage{
insertCol(x, ind, col.names = NULL, data = NA)
}
\arguments{
\item{x}{data.frame}

\item{ind}{Numeric vector. At which index(es) should the new column(s) be inserted.}

\item{col.names}{Character vector. Name(s) of the new column(s).}

\item{data}{Vector, matrix, array or data.frame. Default is NA.}
}
\description{
Insert a column into a data.frame object
}
