% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interleave.R
\name{interleave}
\alias{interleave}
\title{Interleave two or more vectors}
\usage{
interleave(x1, x2, ..., pad = TRUE, padLast = TRUE)
}
\arguments{
\item{x1}{a vector or list}

\item{x2}{a vector or list}

\item{...}{Optional. More vector(s) or list(s)}

\item{pad}{logical: Whether to add NA to shorter object(s). Default TRUE.}

\item{padLast}{logical: When pad is set to TRUE, whether to add NA at the end (TRUE) or the beginning (FALSE). When pad is set to FALSE, whether to cut at the end (TRUE) or beginning (FALSE) Default TRUE.}
}
\value{
same class as input
}
\description{
Interleave two or more vectors
}
\examples{
\dontrun{
}
}
