% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep.signif.R
\name{keep.signif}
\alias{keep.signif}
\title{Title}
\usage{
keep.signif(x, threshold = 0.05)
}
\arguments{
\item{x}{numeric vector between 0 and 1}

\item{threshold}{Default is 0.05}
}
\description{
Title
}
