% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang2str.R
\name{lang2str}
\alias{lang2str}
\title{Convert language to a string}
\usage{
lang2str(x, env = parent.frame(1L))
}
\arguments{
\item{x}{language}
}
\value{
a character vector of length 1
}
\description{
Convert language to a string
}
\examples{
\dontrun{
lang2str(list())
}

}
