% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysisFunctions.R
\name{makeANOVAresultsDF}
\alias{makeANOVAresultsDF}
\title{Put ANOVA results in a data.frame object}
\usage{
makeANOVAresultsDF(data, x, r)
}
\arguments{
\item{data}{data frame object}

\item{x}{vector to use as the x variable in the lme formula}

\item{r}{vector to use as the random variable in the lme formula}
}
\value{
a data frame with the p- and F- value for each column of the data
}
\description{
Put ANOVA results in a data.frame object
}
