% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysisFunctions.R
\name{makeDiversityDF}
\alias{makeDiversityDF}
\title{Calculates shannon diversity, richness and evenness}
\usage{
makeDiversityDF(df)
}
\arguments{
\item{df}{data frame object}
}
\value{
a data frame with the calculated Shannon diversity index, species richness and evenness
}
\description{
Calculates shannon diversity, richness and evenness
}
