% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedList.R
\name{namedList}
\alias{namedList}
\title{Making a named list}
\usage{
namedList(..., evalCall = FALSE, sorted = FALSE)
}
\arguments{
\item{...}{the variables to be include in the list (i.e. \code{a, b, c}) or a call resulting in a list of the variable's name(s) (i.e \code{ls(pattern="^[a-c]$")})}

\item{evalCall}{whether to eval the call made in the ... argument. Default is FALSE.}

\item{sorted}{whether to sort the resulting list by it's names}
}
\value{
A named list from the name(s) and value(s) of the given variable(s)
}
\description{
Making a named list
}
