% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameself.R
\name{nameself}
\alias{nameself}
\title{Name a vector with the its values}
\usage{
nameself(x)
}
\arguments{
\item{x}{vector}
}
\value{
named vector
}
\description{
Name a vector with the its values
}
\examples{
nameself(letters)
}
