% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_list.R
\name{pad_list}
\alias{pad_list}
\title{Pad a vector or list of vector with NAs to a given length.}
\usage{
pad_list(x, length.out = NULL, pad = NA)
}
\arguments{
\item{x}{A list of vector to pad.}

\item{length.out}{integer: The length of vector after padding. Optional. If missing, takes the length of the longer list item.}

\item{pad}{The value to fill the vector(s) with. Default is NA.}
}
\value{
a list or vector depending on input
}
\description{
Pad a vector or list of vector with NAs to a given length.
}
\examples{
\dontrun{
li <- list(a = 1)
}
}
