% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_empty.R
\name{rm_empty}
\alias{rm_empty}
\title{Remove empty elements from a vector or list}
\usage{
rm_empty(x)
}
\arguments{
\item{x}{A vector or list.}
}
\value{
The vector or list without the empty elements (length 0 or empty strings).
}
\description{
Remove empty elements from a vector or list
}
\examples{
\dontrun{
x <- list(0, NULL, "a", NA, "")
rm_empty(x)
}
}
