% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_file_into_oneFunFiles.R
\name{split_file_into_oneFunFiles}
\alias{split_file_into_oneFunFiles}
\title{Split your file(s) containing multiple functions into multiple files containing one function.}
\usage{
split_file_into_oneFunFiles(
  filePath,
  outDir = NULL,
  keepComments = TRUE,
  recursive = TRUE
)
}
\arguments{
\item{filePath}{(character) The path(s) of the directory or file(s) you want to split.}

\item{outDir}{(character) Optional. A directory path where to write the new files. If missing, place them in the filePath directory. Can be the same length as filePath since it will be mapp(ly)ed.}

\item{keepComments}{(logical) Whether to keep the comments before the functions and/or at the end of the file. Default is \code{TRUE}.}

\item{recursive}{(logical) When filePath is a directory, should the file be listed recursively (with \code{list.files}). Default is \code{TRUE}.}
}
\description{
Split your file(s) containing multiple functions into multiple files containing one function.
}
