% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{stack}
\alias{stack}
\title{Stack Vectors from a Data Frame or List}
\usage{
stack(
  x,
  keepNameOrder = FALSE,
  sorted = FALSE,
  keepFactors = FALSE,
  select = NULL
)
}
\description{
Stack Vectors from a Data Frame or List
}
