% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str2lines.R
\name{str2lines}
\alias{str2lines}
\title{Add line breaks to a string at a specified width.}
\usage{
str2lines(string, width = 80)
}
\arguments{
\item{string}{(character) A character vector of length 1.}

\item{width}{(integer|numeric) The number of character per line.}
}
\value{
A character vector of length 1 with the added line breaks.
}
\description{
Add line breaks to a string at a specified width.
}
