% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_helpers.R
\name{str_wrap}
\alias{str_wrap}
\title{Add line breaks to a string at a specified width.}
\usage{
str_wrap(string, width = 80)
}
\arguments{
\item{string}{(character) A character vector of length 1.}

\item{width}{(integer|numeric) The number of character per line.}
}
\value{
A character vector of length 1 with the added line breaks.
}
\description{
Add line breaks to a string at a specified width.
}
