% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toStr.R
\name{toStr}
\alias{toStr}
\title{Convert an R Object to a Character String}
\usage{
toStr(x, and = TRUE, oxford_comma = FALSE)
}
\arguments{
\item{x}{The object to be converted.}

\item{and}{(logical)  Whether to use "and" between the two last item. Default TRUE.}

\item{oxford_comma}{(logical) Whether to use an oxford comma. Default FALSE. Ignored if `and` set to FALSE.}
}
\value{
A character string.
}
\description{
Similar to \link[base]{toString} but with additionnal parameters for more flexibility.
}
\examples{
toStr(letters[1:3])
}
