% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Value(s) validation using partial matching}
\usage{
validate(x, choices, name, several.ok = TRUE)
}
\arguments{
\item{x}{The value(s) to validate}

\item{choices}{A vector of candidate values}

\item{name}{The name of the value being validated (used in error messages)}

\item{several.ok}{(logical) Whether x can have more than one value. Default TRUE.}
}
\description{
Value(s) validation using partial matching
}
\examples{
\dontrun{
 validate(x = "bird", choices = c("dog", "cat"), name = "pet")
}

}
