% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_boolean}
\alias{validate_boolean}
\title{Boolean value(s) validation}
\usage{
validate_boolean(x, name = lang2str(x), keepMissing = FALSE)
}
\arguments{
\item{x}{(character | numeric | integer | logical) A vector.}

\item{name}{(character) The name of the vector (used in error messages). Optional.}

\item{vectorize}{(logical) Whether to vectorize the function when x has length > 1 (TRUE; Default)}

\item{keepNA}{(logical) Whether to keep NA values (TRUE) or coerced them to false (FALSE; Default)}
}
\value{
A logical vector
}
\description{
Check that each element of x is either TRUE or FALSE. Works with vectors containing logical values, and/or 0s and 1s, and/or "true", "t", "false", "f" and "na" strings (regardless of case). Coerce the non logical values to their logical equivalent. If keepMissing is FALSE (default), missing values will be coerced to FALSE.
}
