% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varpart2.R
\name{varpart2}
\alias{varpart2}
\title{Partition the Variation of Community Matrix by 2 or more Explanatory Models}
\usage{
varpart2(
  Y,
  X,
  ...,
  data = parent.frame(),
  chisquare = FALSE,
  transfo = NULL,
  scale = FALSE,
  add = FALSE,
  sqrt.dist = FALSE,
  permutations = 1000L
)
}
\arguments{
\item{Y}{Data frame or matrix containing the response data table or dissimilarity structure inheriting from dist. In community ecology, that table is often a site-by-species table or a dissimilarity object.}

\item{X}{An explanatory model defined by a one-sided formula (~ X).}

\item{...}{One or more other explanatory models defined by a one-sided formula.}

\item{data}{The data frame with the variables used in the formulae in X.}

\item{chisquare}{Partition Chi-square or the inertia of Correspondence Analysis (cca).}

\item{transfo}{Transformation for Y (community data) using decostand. All alternatives in decostand can be used, and those preserving Euclidean metric include "hellinger", "chi.square", "total", "norm". Ignored if Y are dissimilarities.}

\item{scale}{Should the columns of Y be standardized to unit variance. Ignored if Y are dissimilarities.}

\item{add}{Add a constant to the non-diagonal values to euclidify dissimilarities (see wcmdscale for details). Choice "lingoes" (or TRUE) use the recommended method of Legendre & Anderson (1999: “method 1”) and "cailliez" uses their “method 2”. The argument has an effect only when Y are dissimilarities.}

\item{sqrt.dist}{Take square root of dissimilarities. This often euclidifies dissimilarities. NB., the argument name cannot be abbreviated. The argument has an effect only when Y are dissimilarities.}

\item{permutations}{If chisquare = TRUE, the adjusted R^2 is estimated by permutations, and this paramater can be a list of control values for the permutations as returned by the function how, or the number of permutations required, or a permutation matrix where each row gives the permuted indices. (Default is 1000L)}
}
\value{
A list
}
\description{
Partition the Variation of Community Matrix by 2 or more Explanatory Models
}
\note{
Based on the \code{\link[vegan]{varpart}} function and modified to allow more explanatory models.
}
\examples{
\dontrun{
mod <- varpart(mite, mite.env, mite.pcnm, transfo="hel")
mod
summary(mod)
}
}
