% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2str.R
\name{vec2str}
\alias{vec2str}
\title{Convert a vector to a string (for assignment)}
\usage{
vec2str(vec)
}
\arguments{
\item{vec}{a vector to convert}
}
\value{
a string version of the vertoc (ie: 'c("elt1", "elt2", ...)')
}
\description{
Convert a vector to a string (for assignment)
}
\examples{
\dontrun{
grep("^A", names(islands), value = TRUE) |> vec2str(vec = _) |>
  cat()
grep("^A", names(islands), value = TRUE) |> vec2str(vec = _) |>
  clipr::write_clip()
}
}
