% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2str.R
\name{vec2strI}
\alias{vec2strI}
\title{Convert a line of text representing a vector to a string (for assignment)}
\usage{
vec2strI(vec)
}
\arguments{
\item{vec}{a line of text to convert}
}
\value{
inplace replacement of the line to a string
}
\description{
Convert a line of text representing a vector to a string (for assignment)
}
\examples{
\dontrun{
islands_startingWithA <- grep("^A", names(islands), value = TRUE) |> vec2strI(vec = _)
#---
# need a separator or to use alt + enter because ctrl + enter ends on the
# next line and so the range start there, so if the function is there too
# it will act there instead
grep("^A", names(islands), value = TRUE) |> vec2strI(vec = _)
}
}
