% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_dir.R
\name{zip_dir}
\alias{zip_dir}
\title{Use zip from a directory}
\usage{
zip_dir(dirpath, zipname, pattern = NULL, ...)
}
\arguments{
\item{dirpath}{(character vector of length 1) The directory path}

\item{zipname}{(character vector of length 1) The name to give to the zipped file}

\item{pattern}{(character vector of length 1) The pattern to select the files in the directory}

\item{...}{Other parameters to pass to \code{list.files}}
}
\value{
zipname
}
\description{
Use zip from a directory
}
