% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.fun_methods.R
\name{as.fun}
\alias{as.fun}
\title{Convert Formula to Function}
\usage{
as.fun(.body)
}
\arguments{
\item{.body}{either a right sided formula (i.e ~ x==1) or a string describing the wanted function body}
}
\value{
a function
}
\description{
Convert Formula to Function
}
\examples{
data("iris")

as.fun(~ .[.$Species \%in\%  c("setosa","virginica") & .$Petal.Width>0.2,])(iris)
}
