% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toStr.R
\name{toStr}
\alias{toStr}
\title{Convert a vector to a character string}
\usage{
toStr(x, join_word = "and", quote = FALSE, oxford_comma = FALSE)
}
\arguments{
\item{x}{A vector to be converted.}

\item{join_word}{(character) The word to use between the two last item. Default is 'and'. Set to NULL to omit.}

\item{quote}{(logical) Whether to add single quotes around each word.}

\item{oxford_comma}{(logical) Whether to use an oxford comma. Default FALSE. Ignored if \code{and} set to FALSE.}
}
\value{
A character string.
}
\description{
Similar to \link[base:toString]{toString} but with additional parameters for more flexibility.
}
\examples{
toStr(letters[1:3])
toStr(1:5)
}
